feat(i3): setup keyboard variant
The keyboard variant (intl with dead keys) is required to use quotation
as a modifier, plus the letter to apply the modification to.
There's a way to add a shortcut, but for now it is disabled (it is
commented)

This disabled the configuration of the xmodmap that maps CapsLock to
Ctrl. It was required to relaunch the command. This modification was
also added in this commit.